requirements.txt
aiohttp==3.8.6
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
anyio==4.2.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_224434dqzl/croot/attrs_1695717839274/work
beautifulsoup4==4.12.3
bilibili-api-python==16.1.1
Brotli==1.1.0
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_35eq66o3mo/croot/certifi_1700501684871/work/certifi
cffi @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_b4nang6w_y/croot/cffi_1700254307954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
colorama==0.4.6
cryptography @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_b1p0q5vizk/croot/cryptography_1702070293829/work
frozenlist @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_68w5bkvw82/croot/frozenlist_1698702578544/work
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna @ file:///Users/cbousseau/work/recipes/ci_py311/idna_1677906072337/work
lxml==4.9.4
multidict @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_10voz9m15i/croot/multidict_1701096890858/work
Pillow==10.0.1
pyasn1==0.5.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodomex==3.19.1
pyOpenSSL @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_b8whqav6qm/croot/pyopenssl_1690223428943/work
pypng==0.20220715.0
PySocks @ file:///Users/cbousseau/work/recipes/ci_py311/pysocks_1677906386870/work
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
qrcode-terminal==0.8
requests @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_54zi68h2nb/croot/requests_1690400233316/work
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tqdm==4.66.1
typing_extensions==4.9.0
tzlocal==5.2
urllib3 @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_068obtb882/croot/urllib3_1698257558009/work
yarl @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a8s46tbbn5/croot/yarl_1701105147904/work
